dom/asmjscache/test/test_slow.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=854209
-->
<head>
  <meta charset="utf-8">
  <title>asm.js browser tests</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=854209">asm.js browser tests</a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test"></pre>

  <script>
  var jsFuns = SpecialPowers.Cu.getJSTestingFunctions();
  ok(jsFuns.isAsmJSCompilationAvailable(), "asm.js compilation is available");
  </script>

  <script src="http://mochi.test:8888/tests/dom/asmjscache/test/file_slow.js"></script>

  <script>
  var w = new Worker('http://mochi.test:8888/tests/dom/asmjscache/test/file_slow.js');
  w.onmessage = function(e) {
      ok(e.data === "ok", "Worker asm.js tests");
      SimpleTest.finish();
  }

  SimpleTest.waitForExplicitFinish();
  </script>

</body>
</html>
